# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IMR International API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone?: string;
  createdAt?: string;
}

export interface AuthResponse {
  user: UserProfile;
  message: string;
}

export interface EnquiryInput {
  name: string;
  email: string;
  phone: string;
  program: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}