# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IMR International API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface AuthResponse {
  user: UserProfile;
  message: string;
}