# lib/api-zod/src/generated/types/enquiryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IMR International API
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryInput {
  name: string;
  email: string;
  phone: string;
  program: string;
  message?: string;
}